new map and cleaned up cleanup
Made a map (and find/create function) for models
Added new ModelBuffers struct to hold the vertex and index buffers for a model so this can be shared between two model instances.
Changed model destruction to happen in the destructor instead of via an explicit destroy function. There's no way for anybody other than the shared_ptr to know when everybody is done with the buffers to destroy them.