Fixed some causes of blueness
CEF was providing BGR textures that I assumed were RGB. It's now red instead of blue. Made the DXGI code a little more flexible in the process.

Fixed the skybox not being rendered when creating the lighting lookup textures. The papermill skybox is now loaded and rendered properly into these these textures. That makes lighting not a solid blue color across the board.