Unique IDs per client
Switch from AvServerImpl being the actual AvServer::Server object to having an AvServer::Server subclass per connection. This allows each connection to have a unique ID and pass those along to the "real" server on each request.

I'm not using this to clean up things when clients disconnect yet, but that'll come next.