Laid some groundwork for transforms happening in C++
Some small moves in the general direction of javascript pushing unmultiplied transforms into C++. Ideally this will let the traversal code only run when a scene graph change happens.

A stub of a transform class is included in this, but it's barely anything so far.

"shelving" this to clear the desk for websocket work.