Standard hook and more highlighting
Added AvStandardHook, which uses a standard model and highlight behavior. It can just be positioned as part of whatever gadget to add a hook to that gadget.

Added a standard hook to the default hand.

Added plumbing to let hooks know when grabbing starts and stops. This lets them normally be hidden, but show themselves when something might want to drop on them.