Moved aardvark-shared and aardvark-react back into main repo
There was way too much friction involved in switching from public dependencies to internal dependencies constantly. Not being able to go-to-definition or step into functions in the internal dependencies was also a PITA. And so was the need to publish the internal dependencies, then change back to public dependencies on the main project, then build that and publish some more.

Instead of loading @aardvarkxr/aardvark-react and @aardvarkxr/aardvark-shared through NPM, they're now aliases in webpack and tsconfig. Everything that builds out of the main repo will bake in these dependencies.

I'll still need to publish the packages publically when it's time for a release, so those directories retain their package.json and related files. Those just aren't used for the internal references.