Fixed up action state updates per hand/gadget
On the gadget side these were treated as gadget-wide, so now they are on the traverser side too. Traversers keep track of which hands a gadget has an interest in and keeps those action states up to date. When a gadget changes the hands it's interested in, it will get action state updates for the new hand and lose them for the old hand.