Fixed overlay grabby grabbers
Grabbers were reporting themselves as "in range" of a grabber that was being grabbed already. That caused the grabber's state to change to "in range" instead of Grabbed, which messed stuff up. In particular it messed up the control panel, but anything with a distinct grabbed state would have had the problem.

For now we can't transfer from one hand to another as a result #62