Added hysteresis to hooks
Hooks now have an inner and outer volume with a scale that defines the difference. The scale defaults to 1.5, but can be whatever. (Scales less than 1 don't really make much sense and will cause annoying flickering.)

A grabbable that's being grabbed starts having affinity for a hook when it enters the inner (smaller) volume and stops having affinity when it exits the outer (larger) volume. That means you can move a gadget close to your hand, to start hooking it and them move it around a larger area to find the right spot to drop it.