Added new transform types
AvHeadFacingTransform points the positive X axis of itself at the user's head and points the Y axis at "up", more or less. The rotation of the head facing transform's parent is ignored.

AvParentTransform adopts the transform of the specified endpoint address as its transform. This is useful for parenting a node to a node in some other gaget, for instance a grabbable, grabber, or hook.