Added interfaces list to hooks and grabbables
The hook and the grabbable must have at least one matching interface for the hook to apply to that grabbable. If not, it's as though the hook doesn't exist. The default interface if none are specified is "aardvark-gadget@1" in both cases, which retains existing functionality for older gadgets.

Changed the whiteboard marker to have color picker and surface drawing interfaces. Changed the paint buckets to have a color picker interface.