Added composed entities
A composed entity is one that takes any number of concrete components as properties and then makes a single interface entity that uses all of them. Or at least the theory. It works well for the single MoveableComponent I've written on one entity.

Converted the demo ball over to use a MoveableComponent and an AvComposedEntity.