Added networking to the whiteboard
Remotes have a list of IPFS URLs that they get from master, and that's kept up to date as more strokes are added. The just send those down in their scene graph.

When a remote experiences a new stroke it sends it to master for processing and then discards it.

When master gets a new stroke, including from a remote, it makes a model out of that stroke and adds it to IPFS. Then it updates its own stroke URL list and all remotes with the new stroke.

Also fixed a failure that happened with strokes that only have one point in them. Those are just spheres now.