Removed screen_capture_lite from the project
I switched over to using hand-rolled code. The total number of lines of code I was usinging from the library was small, and the way it handles threading didn't really fit with what I needed.

The actual use of the library went away a few days ago. This just gets right of the dependency.