Added grabbable style to AvStandardGrabbable
Rolled AvNetworkGrabbable prototype into AvStandardGrabbable by way a new "style" prop. this is one of Gadget, LocalItem, or NetworkedItem. Gadget is the top level grabbable of a gadget. LocalItem is a non-network-grabbable inside the gadget. NetworkedItem is a networked grabbable inside the gadget.

Fixed up all the AvStandardGrabbable references in the core codebase to use the new style.

Removed advertiseGadgetInfo and useInitialParent because they're both implied by being the Gadget style now.