Interface debugging help
Squelched some events that are really just spam when debugging. The exist to allow gadgets to wait on the event being processed, but aren't helpful when debugging multiple gadgets worth of messages.

Removed the message object from specifically decoded messages. If those messages want a field in the output they can put it in the output.