Added support for "overlay only" models
These models will only render over the parts of the view where something else from Aardvark is being rendered.

Changed the standard rendering to write to the stencil buffer, and then for the overlay-only pass it requires that the stencil buffer be set.

This isn't working yet for transparent geometry. I need to add a pass where that draws only into stencil before the opaque overlay-only geometry and then does it normal pass later to draw into the color buffer.