Fixed skinned meshes
Each node was making a copy of its skin based on the template model, rather than using the copy of the skin that was being created at the model level. And the one at the model level was the only one that had its internal joint node pointers fixed up.

Commented out the joint index fixup for now. That makes no difference for RiggedSimple.glb. If it's still needed for the hands I can put it back.

Fixes #171