Switched poking to be on intersections and not require a button press
This is happening through aardvark-panel@2 interfaces, so it will only work on new gadgets.  Old poking with the "grab" button will still happen on old panels and grab buttons. That's ok for the buttons, but it's pretty awkward for the panels. Does that matter or should they update once this is out and become panel v2?