Fixed early shutdown bug
A landing page was running and it connected to the server before any of the actual gadgets. But the server for that landing page wasn't running, so loading the gadget manifest failed, which caused the server to close that connection, which caused the server to exit because its last connection closed. Now it requires at least one successful connection before that "shut down on last disconnect" behavior is enabled.