Fixed lost grabs
With the old input code, there was a lost grab state to manage grab-is-held-while-grabbable-is-lost state. In the new input code, the input system itself will have that action set deactivated and handles "ignore initial press" so we can just clean up our state.  So I fixed the lost grab state by deleting it.